ovip-0007: make signature verification portable
Before, a verifier needed to deserialize the JSON
in order to extract the signature and then re-serialize
the `Content` back into JSON in order to verify the
signature. Because JSON lacks of a canonical form the
re-serialization into JSON is implementation defined
and therefore not portable.

To work around this limitation the signature is removed
from the JSON document and prepanded to the encoded
JSON bytes. That way signature verification can be
performed without deserialization.